requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
# Open source alternatives
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
langchain==0.0.352
langchain-community==0.0.10
pinecone-client==3.0.0
langchain-pinecone==0.0.3
requests==2.31.0
pypdf2==3.0.1
python-docx==1.1.0
aiofiles==23.2.0
email-validator==2.1.0
# Optional: Keep OpenAI as fallback (commented out)
# openai==1.3.7
# langchain-openai==0.0.2
